+++ backend/reviews/infrastructures/repositories/SbReviewRepository.ts
@@ -1,6 +1,7 @@
 import { supabase } from '@/lib/supabase';
 import { ReviewEntity } from '@/backend/reviews/domains/entities/review';
 import { IReviewRepository } from '@/backend/reviews/domains/repositories/ReviewRepository';
+import { ReviewMapper } from '@/backend/reviews/infrastructures/mappers/ReviewMapper';
 
 export class SbReviewRepository implements IReviewRepository {
   // helpId로 리뷰 리스트 조회
@@ -14,7 +15,7 @@ export class SbReviewRepository implements IReviewRepository {
     if (error) throw error;
     if (!data) return [];
 
-    return data.map((row: any) => this.toEntity(row));
+    return data.map((row: any) => ReviewMapper.toEntity(row));
   }
 
   // 리뷰 id로 단일 리뷰 상세 조회
@@ -28,7 +29,7 @@ export class SbReviewRepository implements IReviewRepository {
     if (error) throw error;
     if (!data) return null;
 
-    return this.toEntity(data);
+    return ReviewMapper.toEntity(data);
   }
 
   // 리뷰 생성
@@ -48,19 +49,7 @@ export class SbReviewRepository implements IReviewRepository {
     if (error) throw error;
     if (!data) throw new Error('리뷰 생성 실패');
 
-    return this.toEntity(data);
+    return ReviewMapper.toEntity(data);
   }
 
-  // DB row → Entity 변환
-  private toEntity(row: any): ReviewEntity {
-    return new ReviewEntity(
-      row.id,
-      row.help_id,
-      row.writer_id,
-      row.receiver_id,
-      row.rating,
-      row.text,
-      new Date(row.created_at)
-    );
-  }
 }
